# app/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "worth-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/dataset/files/{name}": {
      "get": {
        "summary": "Dataset File",
        "operationId": "dataset_file_api_dataset_files__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveredFile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dataset/raw/{path}": {
      "get": {
        "summary": "Dataset Raw",
        "operationId": "dataset_raw_api_dataset_raw__path__get",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "download",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Download"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/fees/localities": {
      "get": {
        "summary": "Fee Localities",
        "operationId": "fee_localities_api_fees_localities_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "today",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Locality"
                  },
                  "title": "Response Fee Localities Api Fees Localities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fees/price": {
      "post": {
        "summary": "Fee Price",
        "operationId": "fee_price_api_fees_price_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Derivation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DeliveredFile": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/TableFile"
          },
          {
            "$ref": "#/components/schemas/NoteBundle"
          },
          {
            "$ref": "#/components/schemas/EdiBundle"
          }
        ]
      },
...
      "Derivation": {
        "properties": {
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "const": "USD",
            "title": "Currency",
            "default": "USD"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "modifier": {
            "type": "string",
            "title": "Modifier"
          },
          "modifiers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modifiers"
          },
          "locality": {
            "type": "string",
            "title": "Locality"
          },
          "localityName": {
            "type": "string",
            "title": "Localityname"
          },
          "placeOfService": {
            "type": "string",
            "title": "Placeofservice"
          },
          "ruleYear": {
            "type": "integer",
            "title": "Ruleyear"
          },
          "quarter": {
            "type": "integer",
            "title": "Quarter"
          },
          "paymentBasis": {
            "type": "string",
            "title": "Paymentbasis"
          },
          "inputs": {
            "$ref": "#/components/schemas/DerivationInputs"
          },
          "adjustedRvuTotal": {
            "type": "string",
            "title": "Adjustedrvutotal"
          },
          "trace": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trace"
          },
          "source": {
            "$ref": "#/components/schemas/DerivationSource"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount",
          "code",
          "modifier",
          "modifiers",
          "locality",
          "localityName",
          "placeOfService",
          "ruleYear",
          "quarter",
          "paymentBasis",
          "inputs",
          "adjustedRvuTotal",
          "trace",
          "source"
        ],
        "title": "Derivation",
        "description": "A priced answer with its arithmetic and its receipts. Mirrors ``worth-fees price --json``."
      },
      "DerivationInputs": {
        "properties": {
          "workRvu": {
            "type": "string",
            "title": "Workrvu"
          },
          "peRvu": {
            "type": "string",
            "title": "Pervu"
          },
          "mpRvu": {
            "type": "string",
            "title": "Mprvu"
          },
          "workGpci": {
            "type": "string",
            "title": "Workgpci"
          },
          "peGpci": {
            "type": "string",
            "title": "Pegpci"
          },
          "mpGpci": {
            "type": "string",
            "title": "Mpgpci"
          },
          "conversionFactor": {
            "type": "string",
            "title": "Conversionfactor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workRvu",
          "peRvu",
          "mpRvu",
          "workGpci",
          "peGpci",
          "mpGpci",
          "conversionFactor"
        ],
        "title": "DerivationInputs",
        "description": "Every input as a decimal string. Strings, because a float is not money."
      },
      "DerivationSource": {
        "properties": {
          "release": {
            "type": "string",
            "title": "Release"
          },
          "releaseDate": {
            "type": "string",
            "title": "Releasedate"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "release",
          "releaseDate",
          "files"
        ],
        "title": "DerivationSource"
      },
...
      "EdiBundle": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "edi",
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "blurb": {
            "type": "string",
            "title": "Blurb"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/EdiFile"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "name",
          "blurb",
          "count",
          "bytes",
          "files"
        ],
        "title": "EdiBundle"
      },
      "EdiFile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "claims": {
            "type": "integer",
            "title": "Claims"
          },
          "lines": {
            "type": "integer",
            "title": "Lines"
          },
          "segments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Segments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "bytes",
          "claims",
          "lines",
          "segments"
        ],
        "title": "EdiFile"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Locality": {
        "properties": {
          "locality": {
            "type": "string",
            "title": "Locality"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mac": {
            "type": "string",
            "title": "Mac"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "locality",
          "name",
          "mac"
        ],
        "title": "Locality"
      },
...
      "NoteBundle": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "notes",
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "blurb": {
            "type": "string",
            "title": "Blurb"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/NoteFile"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "name",
          "blurb",
          "count",
          "bytes",
          "files"
        ],
        "title": "NoteBundle"
      },
...
      "NoteFile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "encounter": {
            "type": "string",
            "title": "Encounter"
          },
          "noteType": {
            "type": "string",
            "title": "Notetype"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "encounter",
          "noteType",
          "author",
          "sha256",
          "bytes",
          "text"
        ],
        "title": "NoteFile"
      },
...
      "PriceRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "locality": {
            "type": "string",
            "title": "Locality"
          },
          "date": {
            "type": "string",
            "title": "Date",
            "default": "today"
          },
          "setting": {
            "type": "string",
            "enum": [
              "facility",
              "non-facility"
            ],
            "title": "Setting",
            "default": "non-facility"
          },
          "modifiers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modifiers"
          },
          "paymentBasis": {
            "type": "string",
            "enum": [
              "non-qualifying-apm",
              "qualifying-apm"
            ],
            "title": "Paymentbasis",
            "default": "non-qualifying-apm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "locality"
        ],
        "title": "PriceRequest"
      },
...
      "SourceLink": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "releaseDate": {
            "type": "string",
            "title": "Releasedate"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "sha256",
          "role",
          "releaseDate",
          "note"
        ],
        "title": "SourceLink"
      },
...
      "TableFile": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "table",
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "blurb": {
            "type": "string",
            "title": "Blurb"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "rows": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "name",
          "blurb",
          "bytes",
          "sha256",
          "columns",
          "rows"
        ],
        "title": "TableFile"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}